new file:   .gitignore 	modified:   .vscode/settings.json 	new file:   data.json 	modified:   expense.cpp 	modified:   expense.hpp 	modified:   expenses.cpp 	modified:   expenses.hpp 	deleted:    main 	modified:   main.cpp 	new file:   main.h

+++ expense.hpp
@@ -6,7 +6,6 @@
 #include <variant>
 #include <string>
 #include <vector>
-#include <regex>
 #include <map>
 
 #include <boost/date_time/gregorian/gregorian.hpp>
@@ -18,19 +17,19 @@ class Expense
 private:
     boost_date _date;
     std::string _description;
-    int32_t _amount;
+    double _amount;
 
 public:
     Expense(const boost_date &,
             const std::string &,
-            const int32_t &);
+            const double &);
     ~Expense();
 
     const boost_date get_date() const;
     const std::string get_description() const;
-    const int32_t get_amount() const;
+    const double get_amount() const;
 
     const void set_date(const boost_date &);
     const void set_description(const std::string &);
-    const void set_amount(const int32_t &);
+    const void set_amount(const double &);
 };
